backend/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.0.6
boto3==1.28.5
flask-jwt-extended==4.5.3
pyjwt==2.3.0
sqlalchemy==1.4.31
psycopg2-binary==2.9.6
redis==5.2.1
python-dateutil==2.8.2
pytest==7.0.1
qrcode==8.1.0